backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
idna==3.10
joblib==1.4.2
numpy==1.24.3
pandas==2.2.3
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
thingspeak==1.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
